IGNITE-26371 Get rid of class name in configurer
